separate search context and method
